modificação de observer para tilemapv2
