+++ src/Tileset.cpp
@@ -1,51 +1,59 @@
+#include <stdio.h>
 #include "Tileset.h"
 #include "Error.h"
-#include "Game.h"
-#include "Camera.h"
 
-#define HIGHLIGHT 30
+#define FORBIDDEN_CHAR ">"
+#define MAX_SPRITE_FILENAME_SIZE 100
 
-TileSet::TileSet(int tileWidth, int tileHeight, string file)
-		: tileSet(file, true)
-		, tileWidth(tileWidth)
-		, tileHeight(tileHeight) {
-
-	REPORT_I_WAS_HERE;
-	rows = tileSet.GetHeight()/tileHeight;
-	columns = tileSet.GetWidth()/tileWidth;
-}
-
-void TileSet::Render(unsigned int index, Vec2 pos, bool highlighted) {
-	ASSERT2((unsigned int)rows*columns > index, "\tReceived "<<index<<" but tileSet has size " << (unsigned int)rows*columns );
-	unsigned int desiredLine, desiredColumn;
-	desiredLine = index/columns;
-	desiredColumn = index%columns;
-	SDL_Rect wantedSubSprite;
-
-	wantedSubSprite.x = desiredColumn*tileWidth;
-	wantedSubSprite.y = desiredLine*tileHeight;
-	wantedSubSprite.w = tileWidth;
-	wantedSubSprite.h = tileHeight;
-	Rect destinyRect(pos.x, pos.y, tileWidth, tileHeight);
-	SDL_Rect dst = Camera::WorldToScreen(destinyRect);
-
-	Color color = Color(tileSet.colorMultiplier.r, tileSet.colorMultiplier.g, tileSet.colorMultiplier.b);
-	if(highlighted){
-		color.r = (tileSet.colorMultiplier.r + HIGHLIGHT) > 255 ? 255 : (tileSet.colorMultiplier.r + HIGHLIGHT);
-		color.g = (tileSet.colorMultiplier.g + HIGHLIGHT) > 255 ? 255 : (tileSet.colorMultiplier.g + HIGHLIGHT);
-		color.b = (tileSet.colorMultiplier.b + HIGHLIGHT) > 255 ? 255 : (tileSet.colorMultiplier.b + HIGHLIGHT);
+TileSet::TileSet(std::string fileName, GameObject &associated){
+	FILE *file= fopen(fileName.c_str(), "r");
+	
+	ASSERT2(file != NULL, "Could not open(or find) file "<< fileName);
+	uint vecSize;
+	fscanf(file, " %fx%f %u", &(tileSize.x), &(tileSize.y), &vecSize);
+	sprites.reserve(vecSize);
+	char spriteName[MAX_SPRITE_FILENAME_SIZE];
+	int numFrames;
+	float durationPerFrame;
+	for(; vecSize >0; vecSize--){
+		ASSERT(3 == fscanf(file, " %[^" FORBIDDEN_CHAR "]" FORBIDDEN_CHAR " %d %f", spriteName, &numFrames, &durationPerFrame) );
+		sprites.emplace_back(std::string(spriteName), associated, durationPerFrame, numFrames);
 	}
-	if ( -1 == SDL_SetTextureColorMod( tileSet.GetTexture().get(), color.r, color.g, color.b) ) {
-		CHECK_SDL_ERROR;
+}
+
+void TileSet::Update(float dt){
+	for(uint i=0; i< sprites.size(); i++){
+		sprites[i].Update(dt);
 	}
-	SDL_RenderCopy(Game::GetInstance().GetRenderer(), tileSet.GetTexture().get(), &wantedSubSprite, &dst);
-//	REPORT_I_WAS_HERE;
 }
 
-int TileSet::GetTileHeight(void) {
-	return tileHeight;
+void TileSet::EarlyUpdate(float dt){
+	for(uint i=0; i< sprites.size(); i++){
+		sprites[i].EarlyUpdate(dt);
+	}
 }
 
-int TileSet::GetTileWidth(void) {
-	return tileWidth;
+void TileSet::LateUpdate(float dt){
+	for(uint i=0; i< sprites.size(); i++){
+		sprites[i].LateUpdate(dt);
+	}
 }
+
+void TileSet::Render() const{
+	printf("[WARNING] TileSet::Render(void) should not be called.\n");
+}
+
+void TileSet::Render(int index, Rect pos) const{
+	sprites[index].Render(pos);
+}
+
+uint TileSet::GetSize(void) const{
+	return sprites.size();
+}
+
+Vec2 TileSet::GetTileSize(void) const{
+	return tileSize;
+}
+
+
+
